app/requirements.txt
wheel
MySQL-python==1.2.5
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Cors==2.0.1
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF
Jinja2==2.7.3
SQLAlchemy==1.0.3
Werkzeug==0.10.4
alembic==0.7.5.post2
argparse==1.2.1
itsdangerous==0.24
mock==1.0.1
nose==1.3.6
python-bcrypt==0.3.1
python-dateutil==2.4.2
uWSGI==2.0.10
wsgiref==0.1.2
facebook-sdk
Flask-Admin
Flask-Restless
pandas
Quandl
nose
